More localization refactoring
